style: 修改方法位置
