+++ jdevelops-data/jdevelops-data-jpa/src/main/java/cn/jdevelops/data/jap/dao/JpaBasicsDao.java
@@ -4,7 +4,7 @@ import cn.jdevelops.data.jap.core.JPAUtilExpandCriteria;
 import cn.jdevelops.data.jap.core.criteria.Restrictions;
 import cn.jdevelops.data.jap.enums.FieldName;
 import cn.jdevelops.data.jap.exception.JpaException;
-import cn.jdevelops.data.jap.util.JpaUtils;
+import cn.jdevelops.data.jap.util.ReflectUtils;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.data.repository.NoRepositoryBean;
@@ -58,7 +58,7 @@ public interface JpaBasicsDao<B, ID> extends JpaRepository<B, ID>, JpaSpecificat
     default B updateEntity(B t) throws JpaException {
        try {
            /* 跟根据ID获取需要更新的数据的 原始数据 */
-           B oidCamera = findById((ID) JpaUtils.getFieldValueByName(FieldName.ID.getFieldName(), t))
+           B oidCamera = findById((ID) ReflectUtils.getFieldValueByName(FieldName.ID.getFieldName(), t))
                    .orElse(null);
            /*
             *将新数据中非空字段 克隆到原始数据中 实现更新
@@ -90,7 +90,7 @@ public interface JpaBasicsDao<B, ID> extends JpaRepository<B, ID>, JpaSpecificat
         try {
             /* 跟根据ID获取需要更新的数据的 原始数据 */
             JPAUtilExpandCriteria<B> jpaSelect = new JPAUtilExpandCriteria<>();
-            jpaSelect.add(Restrictions.eq(selectKey, JpaUtils.getFieldValueByName(selectKey, t), false));
+            jpaSelect.add(Restrictions.eq(selectKey, ReflectUtils.getFieldValueByName(selectKey, t), false));
             B oidCamera = findOne(jpaSelect).orElseThrow(() -> new JpaException("更新失败，查询数据为空"));
             /*
              *将新数据中非空字段 克隆到原始数据中 实现更新
